+++ subpackages/lib/source/odood/lib/assembly/spec.d
@@ -79,10 +79,18 @@ struct AssemblySpecAddon {
 
 struct AssemblySpecSource {
 
-    // Name could be used to reffer to this 
+    /// Name could be used to reffer to this
     string name=null;
+
+    /// URL of git repository
     GitURL git_url;
+
+    /// Reference to fetch (usually branch name)
     string git_ref;
+
+    /// Access group: name of access credential to apply to this repo
+    /// Credentials usually provided via environment variablse
+    string access_group=null;
 
     /// Hash
     @property hashString() const {
@@ -92,21 +100,33 @@ struct AssemblySpecSource {
         return sha1Of(res).toHexString();
     }
 
-    private this(GitURL git_url, in string name=null, in string git_ref=null) {
+    private this(GitURL git_url, in string name=null, in string git_ref=null, in string access_group=null) {
         this.git_url = git_url;
         this.name = name;
         this.git_ref = git_ref;
+        this.access_group = access_group;
     }
 
     private this(in Node yaml_node) {
         enforce!OdoodAssemblyException(
-            yaml_node.containsKey("url"),
-            "Invalid spec! Source url not specified");
-        git_url = yaml_node["url"].as!string;
+            yaml_node.containsKey("url") || yaml_node.containsKey("github") || yaml_node.containsKey("oca"),
+            "Invalid spec! Cannot determine url for git source: no 'url' nor 'github' neither 'oca' property specified.");
+        if (yaml_node.containsKey("url"))
+            git_url = yaml_node["url"].as!string;
+        else if (yaml_node.containsKey("github"))
+            git_url = "https://github.com/" ~ yaml_node["github"].as!string;
+        else if (yaml_node.containsKey("oca"))
+            git_url = "https://github.com/oca/" ~ yaml_node["oca"].as!string;
+        else
+            // Should be unreachable, because check at the start of constructor.
+            assert(0, "Invalid spec");
+
         if (yaml_node.containsKey("name"))
             name = yaml_node["name"].as!string;
         if (yaml_node.containsKey("ref"))
             git_ref = yaml_node["ref"].as!string;
+        if (yaml_node.containsKey("access-group"))
+            access_group = yaml_node["access-group"].as!string;
     }
 
     private Node toYAML() const {